chore(occt-wasm): extract wrapKernelExceptions and oc shim into adapterShims
Moves the last two pieces of construction-time helper code out of
occtWasmAdapter.ts:

- wrapKernelExceptions (Proxy wrapper that converts WebAssembly.Exception
  to readable JS Error)
- buildOcShim (the .oc TopoDS_*/gp_Pnt_3/BRepBuilderAPI_MakeEdge_3 facade
  used by legacy tests that reach into adapter.oc)

The constructor body shrinks from ~55 lines to 4 lines. The shim entries
are kept as `function` declarations (not arrow functions) so they
remain callable via `new oc.TopoDS_Solid()` from tests.

Adapter line count: 1956 → 1868 (-88).
